broke out onboarding stack into its own file
